+++ ChangeLog.md
@@ -22,6 +22,7 @@ See docs/process.md for more on how version tagging works.
 ----------------------
 - Legacy support for ancient vendor-prefixed DOM APIs was removed (#27341,
   #27339, #27338, #27340, #27347)
+- mimalloc was updated to 3.4.1. (#27380)
 
 6.0.3 - 07/13/26
 ----------------